refactor(achievement): moved all achievement stuff to separate module
